requirements.txt
beautifulsoup4==4.6.0
numpy==1.15.1
#opencv-python==4.0.0.21
tqdm==4.23.4
Pillow==5.1.0
requests==2.20.0
imgaug